refactor: move `CteWorkTable`, `default_table_source`   a bunch of files out of core (#15316)
* First Iteration

* fix: CI tests

* stable waypoint, documentation pending, memory moce pending

* stable waypoint 2: add document pending, get a heads up pending

* pushing for test and review

* fix:mock in test

* fix:cliipy
