+++ datafusion/catalog/src/lib.rs
@@ -49,6 +49,8 @@ pub use r#async::*;
 pub use schema::*;
 pub use session::*;
 pub use table::*;
+pub mod cte_worktable;
+pub mod default_table_source;
 pub mod stream;
 pub mod streaming;
 pub mod view;
