+++ datafusion/core/src/datasource/memory.rs
@@ -24,7 +24,6 @@ use std::sync::Arc;
 
 use crate::datasource::{TableProvider, TableType};
 use crate::error::Result;
-use crate::execution::context::SessionState;
 use crate::logical_expr::Expr;
 use crate::physical_plan::insert::{DataSink, DataSinkExec};
 use crate::physical_plan::repartition::RepartitionExec;
@@ -129,7 +128,7 @@ impl MemTable {
     pub async fn load(
         t: Arc<dyn TableProvider>,
         output_partitions: Option<usize>,
-        state: &SessionState,
+        state: &dyn Session,
     ) -> Result<Self> {
         let schema = t.schema();
         let constraints = t.constraints();
@@ -267,6 +266,8 @@ impl TableProvider for MemTable {
     /// # Returns
     ///
     /// * A plan that returns the number of rows written.
+    ///
+    /// [`SessionState`]: crate::execution::context::SessionState
     async fn insert_into(
         &self,
         _state: &dyn Session,
